+++ tests/cart/test_cart.py
@@ -2,12 +2,12 @@ import unittest
 from decimal import Decimal
 from typing import List, Union
 
-from cart.cart import Cart
-from cart.cart_metadata import CartMetadata
-from cart.cart_money import CartMoney
-from finance.currency import Currency
-from finance.money import Money
-from item.order_line_item import OrderLineItem
+from cart import Cart
+from cart import CartMetadata
+from cart import CartMoney
+from finance import Currency
+from finance import Money
+from item import OrderLineItem
 from item.order_quantity_unit import OrderQuantityUnit
